# web/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PreviewResponse": {
        "properties": {
          "host": {
            "title": "Host",
            "type": "string"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "host",
          "workspace_id"
        ],
        "title": "PreviewResponse",
        "type": "object"
      },
...
      "RegisterPreviewBody": {
        "properties": {
          "upstream_host": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Upstream Host",
            "type": "string"
          },
          "upstream_port": {
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Upstream Port",
            "type": "integer"
          }
        },
        "required": [
          "upstream_host",
          "upstream_port"
        ],
        "title": "RegisterPreviewBody",
        "type": "object"
      },
...
      "ShareLinkResponse": {
        "properties": {
          "expires_in_s": {
            "title": "Expires In S",
            "type": "integer"
          },
          "token": {
            "title": "Token",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "token",
          "url",
          "expires_in_s"
        ],
        "title": "ShareLinkResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "GAPT control plane",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/workspaces/{workspace_id}/preview": {
      "delete": {
        "operationId": "unregister_preview_api_workspaces__workspace_id__preview_delete",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unregister Preview",
        "tags": [
          "preview"
        ]
      },
      "post": {
        "operationId": "register_preview_api_workspaces__workspace_id__preview_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterPreviewBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Preview",
        "tags": [
          "preview"
        ]
      }
    },
    "/api/workspaces/{workspace_id}/share": {
      "post": {
        "operationId": "mint_share_link_api_workspaces__workspace_id__share_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "ttl_s",
            "required": false,
            "schema": {
              "default": 3600,
              "minimum": 60,
              "title": "Ttl S",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareLinkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mint Share Link",
        "tags": [
          "preview"
        ]
      }
    },
...
  }
}